Move commands to Core Commands
Missed some refactoring on Useful Commands.

Includes some random lint cleanup.

Renames purge to administration.

The Command class is no longer in use, but I put some comments explaining the direction I would like to go with it.

We will need a script that shows the currently implemented commands since they will not be centralized in one file. This can be exposed via an npm script.